requirements.txt
asgiref==3.5.2
boto3==1.26.15
botocore==1.29.15
certifi==2022.12.7
charset-normalizer==2.1.1
dj-database-url==1.2.0
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
imagekitio==3.0.1
jmespath==1.0.1
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.9.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
whitenoise==6.2.0
